Fix minori

+++ backend/rooms.ts
@@ -2,27 +2,27 @@ import express from 'express';
 import { validateTokenAdmin } from './Utilities/authentication';
 import mongoose from 'mongoose';
 import { connUri, dbOptions } from './Database/databaseService';
-import Rooms from "./Database/schemas/rooms";
+import Rooms from './Database/schemas/rooms';
+import Movies from './Database/schemas/movies';
 
 const router = express.Router();
 
 router.get('', async (req: express.Request, res: express.Response) => {
-
   let db = null;
   try {
     db = await mongoose.createConnection(connUri, dbOptions);
-    const RoomMod = db.model("Rooms", Rooms);
-    let rooms = await RoomMod.find().select("-__v").exec() as any;
-    rooms = rooms.map(({ _id, name, rows }: { _id: string, name: string, rows: number[]}) => {
+    const RoomMod = db.model('Rooms', Rooms);
+    let rooms = (await RoomMod.find().select('-__v').exec()) as any;
+    rooms = rooms.map(({ _id, name, rows }: { _id: string; name: string; rows: number[] }) => {
       return {
         _id,
         name,
-        seatsNumber: rows.reduce((a, b) => a + b, 0),
+        seatsNumber: rows.reduce((a, b) => a + b, 0)
       };
-    })
+    });
     res.status(200).json(rooms);
   } catch (err) {
-    res.status(500).json({ error: "Internal server error." });
+    res.status(500).json({ error: 'Internal server error.' });
   } finally {
     db && db.close();
   }
@@ -37,15 +37,15 @@ router.post('', validateTokenAdmin, async (req: express.Request, res: express.Re
     let db = null;
     try {
       db = await mongoose.createConnection(connUri, dbOptions);
-      const RoomMod = db.model("Rooms", Rooms);
+      const RoomMod = db.model('Rooms', Rooms);
       const room = (await RoomMod.create(elem)) as any;
       res.status(201).json({
         id: room._id,
         name: room.name,
-        seats: 0,
+        seats: 0
       });
     } catch (err) {
-      res.status(500).json({ error: "Internal server error." });
+      res.status(500).json({ error: 'Internal server error.' });
     } finally {
       db && db.close();
     }
@@ -59,15 +59,18 @@ router.delete('/:roomId', validateTokenAdmin, async (req: express.Request, res: 
     let db = null;
     try {
       db = await mongoose.createConnection(connUri, dbOptions);
-      const RoomMod = db.model("Rooms", Rooms);
-      //Check film if uses Room
-      const room = await RoomMod.findByIdAndDelete(id);
-      if (!room)
-        throw new Error();
+      const RoomMod = db.model('Rooms', Rooms);
+      const MovieMod = db.model('Movies', Movies);
+      const isUsed = await MovieMod.exists({ plans: { $elemMatch: { roomId: id } } });
+      if (isUsed) res.status(409).json({ error: 'Non puoi eliminare una sala che è stata già utilizzata per proiettare un film!' });
+      else {
+        const room = await RoomMod.findByIdAndDelete(id);
+        if (!room) throw new Error();
+      }
 
-      res.status(200).json({ message: "Sala rimossa con successo!" });
+      res.status(200).json({ message: 'Sala rimossa con successo!' });
     } catch (err) {
-      res.status(404).json({ error: "Sala non trovata!" });
+      res.status(404).json({ error: 'Sala non trovata!' });
     } finally {
       db && db.close();
     }
@@ -80,16 +83,14 @@ router.put('/:roomId', validateTokenAdmin, async (req: express.Request, res: exp
   else {
     let name = req.body.name;
 
-    if (!name)
-      res.status(400).json({ error: 'Bad request' });
+    if (!name) res.status(400).json({ error: 'Bad request' });
     else {
       let db = null;
       try {
         db = await mongoose.createConnection(connUri, dbOptions);
-        const RoomMod = db.model("Rooms", Rooms);
-        const elem = (await RoomMod.findByIdAndUpdate(id, { name }, { new: true, useFindAndModify: false }) as any);
-        if (!elem)
-          throw new Error();
+        const RoomMod = db.model('Rooms', Rooms);
+        const elem = (await RoomMod.findByIdAndUpdate(id, { name }, { new: true, useFindAndModify: false })) as any;
+        if (!elem) throw new Error();
 
         res.status(200).json({ id: elem._id, name: elem.name });
       } catch (err) {
@@ -108,11 +109,11 @@ router.get('/:roomId/rows', async (req: express.Request, res: express.Response) 
   let db = null;
   try {
     db = await mongoose.createConnection(connUri, dbOptions);
-    const RoomsMod = db.model("Rooms", Rooms);
-    const rooms = await RoomsMod.findById(id) as any;
+    const RoomsMod = db.model('Rooms', Rooms);
+    const rooms = (await RoomsMod.findById(id)) as any;
     res.status(200).json(rooms.rows);
   } catch (err) {
-    res.status(500).json({ error: "Internal server error." });
+    res.status(500).json({ error: 'Internal server error.' });
   } finally {
     db && db.close();
   }
@@ -124,21 +125,19 @@ router.post('/:roomId/rows', validateTokenAdmin, async (req: express.Request, re
 
   if (!id && !seatsNumber && isNaN(seatsNumber)) res.status(400).json({ error: 'Bad request' });
   else {
-
     let db = null;
     try {
       db = await mongoose.createConnection(connUri, dbOptions);
-      const RoomMod = db.model("Rooms", Rooms);
-      const room = await RoomMod.findByIdAndUpdate(id, { $push: { rows: seatsNumber }, new: true, useFindAndModify: false }) as any;
-      if (!room)
-        throw new Error();
+      const RoomMod = db.model('Rooms', Rooms);
+      const room = (await RoomMod.findByIdAndUpdate(id, { $push: { rows: seatsNumber }, new: true, useFindAndModify: false })) as any;
+      if (!room) throw new Error();
 
       res.status(201).json({
         id: room._id,
-        seatsNumber: seatsNumber,
+        seatsNumber: seatsNumber
       });
     } catch (err) {
-      res.status(500).json({ error: "Internal server error." });
+      res.status(500).json({ error: 'Internal server error.' });
     } finally {
       db && db.close();
     }
